RSC convert admin custom-domain page
